lesson3-try2
